:white_check_mark: refactoring specs
