+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
-require "bundler/setup"
-require "ruson"
+require 'bundler/setup'
+require 'ruson'
+require 'support/objects'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
